tests/wpt/web-platform-tests/web-animations/interfaces/AnimationEffectTiming/duration.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffectTiming.duration</title>
<link rel="help" href="https://drafts.csswg.org/web-animations/#dom-animationeffecttiming-duration">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(t => {
  const anim = createDiv(t).animate(null);
  assert_equals(anim.effect.timing.duration, 'auto');
}, 'Has the default value \'auto\'');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
  anim.effect.timing.duration = 123.45;
  assert_times_equal(anim.effect.timing.duration, 123.45,
                     'set duration 123.45');
  assert_times_equal(anim.effect.getComputedTiming().duration, 123.45,
                     'getComputedTiming() after set duration 123.45');
}, 'Can be set to a double value');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
  anim.effect.timing.duration = 'auto';
  assert_equals(anim.effect.timing.duration, 'auto', 'set duration \'auto\'');
  assert_equals(anim.effect.getComputedTiming().duration, 0,
                'getComputedTiming() after set duration \'auto\'');
}, 'Can be set to the string \'auto\'');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] }, { duration: 'auto' });
  assert_equals(anim.effect.timing.duration, 'auto', 'set duration \'auto\'');
  assert_equals(anim.effect.getComputedTiming().duration, 0,
                'getComputedTiming() after set duration \'auto\'');
}, 'Can be set to \'auto\' using a dictionary object');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
  anim.effect.timing.duration = Infinity;
  assert_equals(anim.effect.timing.duration, Infinity, 'set duration Infinity');
  assert_equals(anim.effect.getComputedTiming().duration, Infinity,
                'getComputedTiming() after set duration Infinity');
}, 'Can be set to Infinity');

test(t => {
  const div = createDiv(t);
  assert_throws({ name: 'TypeError' }, () => {
    div.animate({ opacity: [ 0, 1 ] }, -1);
  });
}, 'animate() throws when passed a negative number');

test(t => {
  const div = createDiv(t);
  assert_throws({ name: 'TypeError' }, () => {
    div.animate({ opacity: [ 0, 1 ] }, -Infinity);
  });
}, 'animate() throws when passed negative Infinity');

test(t => {
  const div = createDiv(t);
  assert_throws({ name: 'TypeError' }, () => {
    div.animate({ opacity: [ 0, 1 ] }, NaN);
  });
}, 'animate() throws when passed a NaN value');

test(t => {
  const div = createDiv(t);
  assert_throws({ name: 'TypeError' }, () => {
    div.animate({ opacity: [ 0, 1 ] }, { duration: -1 });
  });
}, 'animate() throws when passed a negative number using a dictionary object');

test(t => {
  const div = createDiv(t);
  assert_throws({ name: 'TypeError' }, () => {
    div.animate({ opacity: [ 0, 1 ] }, { duration: -Infinity });
  });
}, 'animate() throws when passed negative Infinity using a dictionary object');

test(t => {
  const div = createDiv(t);
  assert_throws({ name: 'TypeError' }, () => {
    div.animate({ opacity: [ 0, 1 ] }, { duration: NaN });
  });
}, 'animate() throws when passed a NaN value using a dictionary object');

test(t => {
  const div = createDiv(t);
  assert_throws({ name: 'TypeError' }, () => {
    div.animate({ opacity: [ 0, 1 ] }, { duration: 'abc' });
  });
}, 'animate() throws when passed a string other than \'auto\' using a'
   + ' dictionary object');

test(t => {
  const div = createDiv(t);
  assert_throws({ name: 'TypeError' }, () => {
    div.animate({ opacity: [ 0, 1 ] }, { duration: '100' });
  });
}, 'animate() throws when passed a string containing a number using a'
   + ' dictionary object');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
  assert_throws({ name: 'TypeError' }, () => {
    anim.effect.timing.duration = -1;
  });
}, 'Throws when setting a negative number');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
  assert_throws({ name: 'TypeError' }, () => {
    anim.effect.timing.duration = -Infinity;
  });
}, 'Throws when setting negative infinity');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
  assert_throws({ name: 'TypeError' }, () => {
    anim.effect.timing.duration = NaN;
  });
}, 'Throws when setting a NaN value');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
  assert_throws({ name: 'TypeError' }, () => {
    anim.effect.timing.duration = 'abc';
  });
}, 'Throws when setting a string other than \'auto\'');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
  assert_throws({ name: 'TypeError' }, () => {
    anim.effect.timing.duration = '100';
  });
}, 'Throws when setting a string containing a number');

promise_test(t => {
  const anim = createDiv(t).animate(null, 100 * MS_PER_SEC);
  return anim.ready.then(() => {
    const originalStartTime   = anim.startTime;
    const originalCurrentTime = anim.currentTime;
    assert_equals(anim.effect.getComputedTiming().duration, 100 * MS_PER_SEC,
                  'Initial duration should be as set on KeyframeEffect');

    anim.effect.timing.duration = 200 * MS_PER_SEC;
    assert_equals(anim.effect.getComputedTiming().duration, 200 * MS_PER_SEC,
                  'Effect duration should have been updated');
    assert_times_equal(anim.startTime, originalStartTime,
                       'startTime should be unaffected by changing effect ' +
                       'duration');
    assert_times_equal(anim.currentTime, originalCurrentTime,
                       'currentTime should be unaffected by changing effect ' +
                       'duration');
  });
}, 'Extending an effect\'s duration does not change the start or current time');

test(t => {
  const div = createDiv(t);
  const anim = div.animate(null, { duration: 100000, fill: 'both' });
  anim.finish();
  assert_equals(anim.effect.getComputedTiming().progress, 1,
                'progress when animation is finished');
  anim.effect.timing.duration *= 2;
  assert_times_equal(anim.effect.getComputedTiming().progress, 0.5,
                     'progress after doubling the duration');
  anim.effect.timing.duration = 0;
  assert_equals(anim.effect.getComputedTiming().progress, 1,
                'progress after setting duration to zero');
  anim.effect.timing.duration = 'auto';
  assert_equals(anim.effect.getComputedTiming().progress, 1,
                'progress after setting duration to \'auto\'');
}, 'Can be updated while the animation is in progress');

</script>
</body>
